+++ internal/types/order.go
@@ -17,7 +17,7 @@ const (
 
 var (
 	orderTypeString  = "__asc_desc_iasc_idesc_undef"
-	orderTypeIdx     = [...]int{0, 2, 6, 11, 16, 22, 27}
+	orderTypeIdx     = [...]int8{0, 2, 6, 11, 16, 22, 27}
 	orderTypeReverse = map[string]OrderType{}
 )
